+++ extensions/assets-manager/CCAsyncTaskPool.h
